requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
idna==3.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.11
